Use target profiles
This allows for all of the details for a particular target to be
kept separate from the rest of the builder, making it easy to expand
or change the target.

For the most part things should not have different behavior on
different targets, the only exception will be platform-specific
packages that must be held back with apt or differences in the way
the base images are organized (some don't have a boot partition).

+++ stages/01-Baseimage/01-run.sh
@@ -1,22 +1,24 @@
 # Do this to the WORK folder of this stage
 pushd ${STAGE_WORK_DIR}
 
-log "Create 2Gb empty image"
-dd if=/dev/zero of=temp.img bs=1 count=1 seek=2G
+if [[ "${OS}" == "raspbian" ]]; then
+
+    log "Create 2Gb empty image"
+    dd if=/dev/zero of=temp.img bs=1 count=1 seek=2G
 
 
-log "Enlarge the downloaded image by 2Gb"
-cat temp.img >> IMAGE.img
+    log "Enlarge the downloaded image by 2Gb"
+    cat temp.img >> IMAGE.img
 
-log "fdisk magic to enlarge the main partition"
+    log "fdisk magic to enlarge the main partition"
 
-PARTED_OUT=$(parted -s IMAGE.img unit s print)
-ROOT_OFFSET=$(echo "$PARTED_OUT" | grep -e '^ 2'| xargs echo -n \
-    | cut -d" " -f 2 | tr -d s)
+    PARTED_OUT=$(parted -s IMAGE.img unit s print)
+    ROOT_OFFSET=$(echo "$PARTED_OUT" | grep -e "^ ${ROOT_PART}"| xargs echo -n \
+        | cut -d" " -f 2 | tr -d s)
 
-echo "ROOT OFFSET: $ROOT_OFFSET"
+    echo "ROOT OFFSET: $ROOT_OFFSET"
 
-fdisk IMAGE.img <<EOF
+    fdisk IMAGE.img <<EOF
 d
 2
 n
@@ -26,6 +28,7 @@ ${ROOT_OFFSET}
 
 w
 EOF
+fi
 
 # return
 popd
